convolution

+++ 220502_led_disk/Disc.h
@@ -11,7 +11,7 @@ public:
 
   int _pin;
 
-  float confVal;
+  int confVal;
   int beatTime;
 
 private:
